+++ packages/flutter_series/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+* Added support for light themes / changing divider colors.
+* Added comments to flutter_series.
+
 ## 0.1.0
 
 * Initial release